Refactors UserAgents charts in order to reduce the abstraction
Fixes: OX-9135
