+++ src/main/java/sirius/biz/tycho/DashboardBrowserDistributionChart.java
@@ -8,14 +8,21 @@
 
 package sirius.biz.tycho;
 
-import sirius.biz.analytics.charts.explorer.BrowserDistributionTimeSeriesChartFactory;
+import sirius.biz.analytics.charts.explorer.BrowserDistributionTimeSeriesComputer;
+import sirius.biz.analytics.charts.explorer.ChartFactory;
+import sirius.biz.analytics.charts.explorer.ChartObjectResolver;
+import sirius.biz.analytics.charts.explorer.TimeSeriesChartFactory;
+import sirius.biz.analytics.charts.explorer.TimeSeriesComputer;
 import sirius.biz.analytics.events.PageImpressionEvent;
+import sirius.biz.jobs.StandardCategories;
 import sirius.biz.tenants.TenantUserManager;
-import sirius.db.jdbc.SmartQuery;
+import sirius.kernel.commons.Callback;
 import sirius.kernel.di.std.Register;
 import sirius.web.security.Permission;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.function.Consumer;
 
 /**
  * Implements a time series chart for the dashboard, showing the browser distribution.
@@ -24,7 +31,41 @@ import javax.annotation.Nonnull;
  */
 @Register
 @Permission(TenantUserManager.PERMISSION_SYSTEM_TENANT_MEMBER)
-public class DashboardBrowserDistributionChart extends BrowserDistributionTimeSeriesChartFactory<PageImpressionEvent> {
+public class DashboardBrowserDistributionChart extends TimeSeriesChartFactory<Object> {
+    @Nullable
+    @Override
+    protected Class<? extends ChartObjectResolver<Object>> getResolver() {
+        return null;
+    }
+
+    @Override
+    public String getCategory() {
+        return StandardCategories.MISC;
+    }
+
+    @Override
+    protected void collectReferencedCharts(Consumer<Class<? extends ChartFactory<Object>>> referenceChartConsumer) {
+        // intentionally empty -> there are no referenced charts
+    }
+
+    @Override
+    protected void computers(boolean hasComparisonPeriod,
+                             boolean isComparisonPeriod,
+                             Callback<TimeSeriesComputer<Object>> executor) throws Exception {
+        if (isComparisonPeriod) {
+            return;
+        }
+
+        executor.invoke(new BrowserDistributionTimeSeriesComputer<>(PageImpressionEvent.class,
+                                                                    (ignored, query) -> query.eq(PageImpressionEvent.AGGREGATION_URI,
+                                                                                                 "/system/dashboard")));
+    }
+
+    @Override
+    protected boolean stackValues(boolean hasComparisonPeriod) {
+        return true;
+    }
+
     @Nonnull
     @Override
     public String getName() {
@@ -34,15 +75,5 @@ public class DashboardBrowserDistributionChart extends BrowserDistributionTimeSe
     @Override
     public int getPriority() {
         return 901;
-    }
-
-    @Override
-    protected Class<PageImpressionEvent> getEvent() {
-        return PageImpressionEvent.class;
-    }
-
-    @Override
-    protected void modifyQuery(SmartQuery<PageImpressionEvent> query) {
-        query.eq(PageImpressionEvent.AGGREGATION_URI, "/system/dashboard");
     }
 }
